+++ resources/js/pages/users/Permissions.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
   import {computed} from 'vue';
   import {t} from '@craftcms/ui';
-  import {useForm, usePage} from '@inertiajs/vue3';
+  import {router, useForm, usePage} from '@inertiajs/vue3';
   import CpLink from '@/common/components/CpLink.vue';
   import {useAppLayout} from '@/common/composables/useAppLayout';
   import PermissionTree from '@craftcms/ui/vue/CraftPermissionTree.vue';
@@ -87,6 +87,7 @@
         :can-create="props.can.createGroups"
         :error="form.errors.groups"
         v-model="form.groups"
+        @created="router.reload({only: ['groups']})"
       />
     </craft-field-group>